CRQ-3 refactoring sorce code related to chain data, and add rpcserver (mockup for testing)
