+++ requirements.txt
@@ -7,3 +7,8 @@ prometheus-client==0.15.0
 git+https://github.com/bifrost-platform/solidity-contract-configs@0.2.1
 python-dotenv==0.21.0
 jsonpath-ng
+
+jsonrpcserver
+fastapi
+uvicorn
+starlette
